fix(src): adds an options method for compiling default options and changes all interfaces
the new options method is used internally by each interface to generate the options payload for the
right multi method and adds an ApiOption named `concise` that instructs the interfaces whether or
not to return the data object or the full response object. Now, the interfaces only return data
instead of the full reponse object.