feat(*): moves the core data types out of common/types and into common/schema and introduces tsoa
this update touches a lot of files, but the interfaces from here on out should not be changing in
any way. the addition of the server.ts file is an attempt to address #4, using tsoa. testing and
refinements still need to be made on that front. the schema file is also a move in the direction of
having this file be generated by the sanity project biketag-games.