feat(imgur): adds the archiveTag method to the client and includes functionality for Imgur adapter
this update includes changes to the schema to include a field named archivehash on the Game type.
this update also includes several fixes around deleteTag and payloads across other methods.

BREAKING CHANGE: Some of the payloads have changed to not include a tag field, while the payload
itself is a tag compatible type. This impacts the existing updateTagPayload and deleteTagPayload
types.