fix(imgur): fixes a bug in the getTags method for the imgur adapter
this update fixes a faintly hidden bug where the msytery and found images were mixed up for a given
tag. I added a bit more fuzzy logic for determining if a tag is a mystery tag. In the future, could
also add a lookup for the foundLocation and if that ends up empty it's more than likely a mystery
tag.