fix(auth): fixes/implements auth where it wasn't previously integrated
this update makes changes to both the assigning of credentials as well as the requests made using
biketag.request (the default fetcher) which will now negotiate an oauth2 transaction with a biketag
api server. This update includes a new method: fetchCredentials which will obtain authorization (if
it needs to) before making a call to the biketag api server for resource claims. The results of that
method are that the internal configuration is updated with the new credentials.